rename some things
After getting some feedback I decided to rename
ExponentialBackoff to Backoff. I also removed the
word `Default` from in front of the two default backoff
policies the lib provides. This seems to be a cleaner api.

Also, it was recommend I add a link to the docs, so I did!

I would like to thank @acln0 for these suggestions.