refactor It
I decided that the future is now, and that future is that all go
code should be context aware. I have "broken" the api. Now it
always needs a context to be provided.

I also added a convenience method on Backoff called It. Incase
you want to pass one less thing in and you don't like my clever
syntax. This will just call the package level fuction under the
hood.